+++ pytsite/db/_function.py
@@ -67,7 +67,7 @@ def get_collection(name: str) -> _Collection:
         global __client, __database
         __client = None
         __database = None
-        _logger.error('{}. Connection lost.'.format(__name__))
+        _logger.error('Connection lost.', __name__)
         return get_collection(name)
 
 
@@ -80,5 +80,5 @@ def get_collection_names(include_system: bool=False) -> list:
         global __client, __database
         __client = None
         __database = None
-        _logger.error('{}. Connection lost.'.format(__name__))
+        _logger.error('Connection lost.', __name__)
         return get_collection_names(include_system)