+++ pytsite/odm_ui/eps.py
@@ -84,7 +84,7 @@ def post_m_form(args: dict, inp: dict) -> _http.response.Redirect:
         _router.session.add_info(_lang.t('pytsite.odm_ui@operation_successful'))
     except Exception as e:
         _router.session.add_error(str(e))
-        _logger.error(str(e))
+        _logger.error(str(e), __name__)
 
     return _http.response.Redirect(form.redirect)